fix(ci): forge fmt — auto-format generated verifier + vault

+++ contracts/src/ConfidentialTransferVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
-    uint256 constant alphay  = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
-    uint256 constant betax1  = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
-    uint256 constant betax2  = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
-    uint256 constant betay1  = 728736684120689475367658410355586091946600404348752493917377386977690379436;
-    uint256 constant betay2  = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
+    uint256 constant alphax = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
+    uint256 constant alphay = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
+    uint256 constant betax1 = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
+    uint256 constant betax2 = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
+    uint256 constant betay1 = 728736684120689475367658410355586091946600404348752493917377386977690379436;
+    uint256 constant betay2 = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,33 +42,36 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 21444681002309722864389681280833124909709805956711887630453776975612190585276;
     uint256 constant deltay2 = 10605357755521956531358314685111519167727186983396767491013762193531226912565;
 
-    
     uint256 constant IC0x = 12881521001946870271942423304563362607671861355755887612376691046225200347452;
     uint256 constant IC0y = 13908892872099119246277381805329505314570785167283103861742337075819040205261;
-    
+
     uint256 constant IC1x = 20878315419777210855025257558521001530597328810734364895925980024490320548081;
     uint256 constant IC1y = 6471100208598535532837596284084919879045469509124346055743324078698410984011;
-    
+
     uint256 constant IC2x = 20136971203030202293084005647546687607446411499602014529168128085399203876369;
     uint256 constant IC2y = 8497479827257844483812643593832164054425739331845038659152348755082496525822;
-    
+
     uint256 constant IC3x = 1069542059993641090409939146016753019758724216859548792607031865182696136463;
     uint256 constant IC3y = 19781118125336516228374994684344834099776565908904135474663359970764785506852;
-    
+
     uint256 constant IC4x = 17117489404278077335817185413252778131101474975440699072496908460490967585007;
     uint256 constant IC4y = 5977316925346016593089845586900520983262281183924660271176537935854989989798;
-    
+
     uint256 constant IC5x = 10735605210981414254950987548829197294917132783908134974232704069828143861365;
     uint256 constant IC5y = 5882406651420016568674343240821222287219018538313723025866433950934095651019;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[5] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[5] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -76,7 +79,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -111,17 +114,16 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
+
                 g1_mulAccC(_pVk, IC5x, IC5y, calldataload(add(pubSignals, 128)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -147,7 +149,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -164,7 +165,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -174,23 +174,22 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
+
             checkField(calldataload(add(_pubSignals, 128)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
